+++ src/Plugin/PluginManager.php
@@ -6,6 +6,9 @@ use Amp\Promise;
 use Cspray\Labrador\AsyncEvent\Emitter;
 use Auryn\Injector;
 
+use Cspray\Labrador\Exception\Exception;
+use Cspray\Labrador\Exception\InvalidArgumentException;
+use Cspray\Labrador\Exception\InvalidStateException;
 use Ds\Map;
 use Ds\Pair;
 use Ds\Set;
@@ -16,7 +19,7 @@ use Psr\Log\LoggerAwareInterface;
 use Psr\Log\LoggerAwareTrait;
 
 /**
- * De facto Pluggable implementation that manages the lifecycle of Plugins for all out-of-the-box Applications.
+ * The default Pluggable implementation that manages the lifecycle of Plugins for all out-of-the-box Applications.
  *
  * It is HIGHLY recommended that if you implement your own Pluggable interface that you delegate the actual
  * responsibilities for handling the lifecycle of the Plugin to an instance of this object; it is well tested and
@@ -44,6 +47,17 @@ final class PluginManager implements Pluggable, LoggerAwareInterface {
 
     private $pluginsLoaded = false;
 
+    /**
+     * Constructs the PluginManager with dependencies required to be provided to certain Plugin types.
+     *
+     * There are 2 primary reasons for asking for the Injector in this class; the first is that we are required to
+     * construct Plugins from a string and constructing objects in a fashion that all known dependencies are provided
+     * is a natural responsibility of the Injector. The second is that the InjectorAwarePlugin requires that an Injector
+     * be provided during the plugin loading process.
+     *
+     * @param Injector $injector
+     * @param Emitter $emitter
+     */
     public function __construct(Injector $injector, Emitter $emitter) {
         $this->injector = $injector;
         $this->emitter = $emitter;
@@ -62,7 +76,13 @@ final class PluginManager implements Pluggable, LoggerAwareInterface {
         $vector->push(new Pair($pluginHandler, $arguments));
     }
 
+    /**
+     * @param string $plugin
+     * @throws InvalidArgumentException
+     * @throws InvalidStateException
+     */
     public function registerPlugin(string $plugin) : void {
+        /** @var InvalidArgumentException|InvalidStateException $exception */
         if ($exception = $this->guardRegisterPluginIsValid($plugin)) {
             throw $exception;
         }
@@ -70,7 +90,12 @@ final class PluginManager implements Pluggable, LoggerAwareInterface {
         $this->logger->info(sprintf('Registered Plugin "%s".', $plugin));
     }
 
-    private function guardRegisterPluginIsValid(string $plugin) : ?\Throwable {
+    /**
+     * @param string $plugin
+     * @return Exception|null
+     * @throws InvalidArgumentException
+     */
+    private function guardRegisterPluginIsValid(string $plugin) : ?Exception {
         if ($this->hasPluginBeenRegistered($plugin)) {
             return Exceptions::createException(
                 Exceptions::PLUGIN_ERR_HAS_BEEN_REGISTERED,
@@ -153,8 +178,15 @@ final class PluginManager implements Pluggable, LoggerAwareInterface {
         return $this->pluginsLoaded;
     }
 
+    /**
+     * @param string $name
+     * @return Plugin
+     * @throws InvalidArgumentException
+     * @throws InvalidStateException
+     */
     public function getLoadedPlugin(string $name) : Plugin {
         if (!$this->hasPluginBeenRegistered($name)) {
+            /** @var InvalidArgumentException $exception */
             $exception = Exceptions::createException(
                 Exceptions::PLUGIN_ERR_PLUGIN_NOT_FOUND,
                 null,
@@ -165,6 +197,7 @@ final class PluginManager implements Pluggable, LoggerAwareInterface {
 
         $plugin = $this->plugins->get($name);
         if ($plugin === null) {
+            /** @var InvalidStateException $exception */
             $exception = Exceptions::createException(
                 Exceptions::PLUGIN_ERR_INVALID_PLUGIN_ACCESS_PRELOAD,
                 null
@@ -175,8 +208,14 @@ final class PluginManager implements Pluggable, LoggerAwareInterface {
         return $plugin;
     }
 
+    /**
+     * @return Set
+     * @throws InvalidArgumentException
+     * @throws InvalidStateException
+     */
     public function getLoadedPlugins() : Set {
         if (!$this->havePluginsLoaded()) {
+            /** @var InvalidStateException $exception */
             $exception = Exceptions::createException(
                 Exceptions::PLUGIN_ERR_INVALID_PLUGIN_ACCESS_PRELOAD,
                 null
@@ -250,7 +289,13 @@ final class PluginManager implements Pluggable, LoggerAwareInterface {
         });
     }
 
-    private function guardLoadingValidPluginDependency(string $plugin, string $reqPluginName) : ?\Throwable {
+    /**
+     * @param string $plugin
+     * @param string $reqPluginName
+     * @return Exception|null
+     * @throws InvalidArgumentException
+     */
+    private function guardLoadingValidPluginDependency(string $plugin, string $reqPluginName) : ?Exception {
         if ($this->isLoading($reqPluginName)) {
             return Exceptions::createException(
                 Exceptions::PLUGIN_ERR_CIRCULAR_DEPENDENCY,